# Program configuration optimizations