--- a/js/render/draw_background.js
+++ b/js/render/draw_background.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const pixelsToTileUnits = require('../source/pixels_to_tile_units');
-const ProgramConfiguration = require('../data/program_configuration');
 
 const tileSize = 512;
 
@@ -47,10 +46,7 @@ function drawBackground(painter, sourceCache, layer) {
         // Draw filling rectangle.
         if (painter.isOpaquePass !== (color[3] === 1)) return;
 
-        program = painter.useProgram('fill', ProgramConfiguration.createStatic([
-            {name: 'u_color', components: 4},
-            {name: 'u_opacity', components: 1}
-        ]));
+        program = painter.useProgram('fill', painter.basicFillProgramConfiguration);
 
         gl.uniform4fv(program.u_color, color);
         gl.uniform1f(program.u_opacity, opacity);