Along with #3485, this PR brings the rendering performance with the Streets z13 benchmark to pre-DDS levels and closes #3342! 🎉🎉🎉 [Benchmark](http://jsfiddle.net/Mourner/2bzyLLp5/4/) results:

- v0.22: 8.8ms
- v0.25: 17.2ms
- master before #3485: 15.3ms
- master: 11.7ms
- after this PR: 8.3ms

Changes:

- Avoid repeatedly restringifying program configurations for caching.
- Avoid repeatedly recreating the same fill program for clipping masks and background.

Note that the failing build is due to changes in ESLint 3.9.0; fixed up the warnings in the third commit.

👀 @jfirebaugh @lucaswoj @ansis