@mourner 2016-10-30
@jfirebaugh oops, nope, the two render tests for switching paint from default to property function are failing. Investigating.