@mourner 2016-10-30
OK fixed the bug and rebased, merging on green.