@mourner 2016-10-30
For the record, `frame-duration` benchmark results:

**before both PRs**: 9.6 ms, 5% > 16ms
**after**: 7.4 ms, 0% > 16ms