@mourner 2016-10-30
Note that all these optimizations became much more obvious after the ProgramConfiguration #3439 refactor. It made understanding how programs are created and used **much** easier.